HEALTHCARE-FHIR-PIPELINE/requirements.txt
pandas==2.1.4
pyarrow==14.0.1
sqlalchemy==2.0.23
pyodbc==5.0.1
azure-storage-blob==12.19.0
faker==22.0.0
python-dotenv==1.0.0
requests==2.31.0
numpy==1.26.2
openpyxl==3.1.2
tabulate==0.9.0
